#51 💰 feat : 유효성검사, 에러 알림창, 컴포넌트 분리1
